internalize _getInitializedChains and delete chains/index.ts
